+++ lib/llm_cost_tracker/monthly_total.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require "active_record"
-
-module LlmCostTracker
-  class MonthlyTotal < ActiveRecord::Base
-    self.table_name = "llm_cost_tracker_monthly_totals"
-  end
-end