+++ lib/llm_cost_tracker/period_total.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "active_record"
+
+module LlmCostTracker
+  class PeriodTotal < ActiveRecord::Base
+    self.table_name = "llm_cost_tracker_period_totals"
+  end
+end